Third commit
Bootstrap currently not working.

+++ GroupAssignment1/Models/Housing.cs
@@ -4,7 +4,7 @@
     {
         public int HousingId { get; set; }
         public string Name { get; set; } = string.Empty;
-        public decimal Rent { get; set; }
+        public double Rent { get; set; }
         public string? Description { get; set; }
         public string? ImageUrl { get; set; }   
     }
